fix: resolve build errors in host modules and tests
Fix API mismatches and include paths across multiple modules:
- IOModule: Fix mouse method calls (MouseRelease→MouseUp, MousePress→MouseDown)
- MediaModule: Stub MPVController methods not in API
- HTTPModule: Fix include path (HttpClient→HttpModule)
- RuntimeModule: Use public setters for debug flags, stub cliArgs
- WindowModule: Fix WindowManager/Window API calls
- TimerModule: Fix pointer cast with uintptr_t intermediate
- DetectorModule: Fix include path
- HelpModule: Truncate to valid module code
- Various modules: Fix lambda capture lists for valueToString/valueToNumber
- Test files: Fix include paths and linking issues

Project now builds successfully with havel executable at 4.8MB.

+++ src/modules/timer/TimerModule.cpp
@@ -44,7 +44,7 @@ void registerTimerModule(Environment& env, HostContext&) {
             });
             
             // Return timer ID (use pointer address as ID)
-            return HavelValue(reinterpret_cast<double>(timer.get()));
+            return HavelValue(static_cast<double>(reinterpret_cast<uintptr_t>(timer.get())));
         } catch (const std::exception& e) {
             return HavelRuntimeError("Failed to create timer: " + std::string(e.what()));
         }
@@ -75,7 +75,7 @@ void registerTimerModule(Environment& env, HostContext&) {
                      callback->declaration->name->symbol);
             });
             
-            return HavelValue(reinterpret_cast<double>(timer.get()));
+            return HavelValue(static_cast<double>(reinterpret_cast<uintptr_t>(timer.get())));
         } catch (const std::exception& e) {
             return HavelRuntimeError("Failed to create interval: " + std::string(e.what()));
         }
